+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/server/repositories/translations_repository.dart
@@ -1,0 +1,20 @@
+{{> licence.dart }}
+
+import 'dart:convert';
+
+class TranslationsRepository {
+  Map<String, dynamic> getTranslations() => _demoTranslations;
+
+  final Map<String, dynamic> _demoTranslations = jsonDecode('''
+    {
+       "en":{
+          "_ok":"Okay",
+          "login___logIn" : "Login via email"
+       },
+       "bg":{
+          "_ok":"ok",
+          "login___logIn" : "Вход с имейл"
+       }
+    }
+    ''');
+}