+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/server/utils/dependency_injector.dart
@@ -1,0 +1,18 @@
+{{> licence.dart }}
+
+/// Manager class used to keep track of any registered dependencies
+class DependencyInjector {
+  static final Map<Type, dynamic> _dependencies = {};
+
+  /// Registers dependency of given type. If entry of existing type exists, it
+  /// will be overridden
+  T register<T>(T dependency) {
+    _dependencies[dependency.runtimeType] = dependency;
+    return dependency;
+  }
+
+  /// Retrieves dependency of given type
+  T get<T>() {
+    return _dependencies[T] as T;
+  }
+}