+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/base/app/{{project_name}}.dart
@@ -10,8 +10,10 @@ import 'package:flutter_rx_bloc/flutter_rx_bloc.dart';
 import 'package:local_session_timeout/local_session_timeout.dart';{{/enable_pin_code}}
 import 'package:provider/provider.dart'; {{#enable_change_language}}
 import 'package:widget_toolkit/language_picker.dart'; {{/enable_change_language}}
-import '../../l10n/l10n.dart';{{#has_authentication}}{{#analytics}}
-import '../../lib_analytics/blocs/analytics_bloc.dart';{{/analytics}}{{#enable_pin_code}}
+
+import '../../assets.dart';
+import '../../l10n/{{project_name}}_app_i18n.dart';{{#analytics}}
+import '../../lib_analytics/blocs/analytics_bloc.dart';{{/analytics}}{{#has_authentication}}{{#enable_pin_code}}
 import '../../lib_auth/blocs/user_account_bloc.dart';{{/enable_pin_code}}
 import '../../lib_auth/data_sources/remote/interceptors/auth_interceptor.dart';{{/has_authentication}} {{#enable_change_language}}
 import '../../lib_change_language/bloc/change_language_bloc.dart';{{/enable_change_language}}{{#enable_dev_menu}}
@@ -39,8 +41,8 @@ class {{project_name.pascalCase()}} extends StatelessWidget {
     this.config = EnvironmentConfig.production,
 {{#enable_dev_menu}}
     this.createDevMenuInstance,{{/enable_dev_menu}}
-  Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   final EnvironmentConfig config;
 {{#enable_dev_menu}}
@@ -248,7 +250,7 @@ Widget _buildMaterialApp(BuildContext context) => MaterialApp.router(
        theme: {{project_name.pascalCase()}}Theme.buildTheme(DesignSystem.light()),
        darkTheme: {{project_name.pascalCase()}}Theme.buildTheme(DesignSystem.dark()),
        localizationsDelegates: const [
-         I18n.delegate,
+         AppI18n.delegate,
          ...GlobalMaterialLocalizations.delegates,
        ],
        supportedLocales: I18n.supportedLocales,