+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_splash/services/splash_service.dart
@@ -1,17 +1,30 @@
 {{> licence.dart }}
 
 import '../../lib_permissions/services/permissions_service.dart';
+import '../../lib_translations/services/translations_service.dart';
 
 class SplashService {
-  SplashService(this.permissionsService);
+  SplashService(
+    PermissionsService permissionsService,
+    TranslationsService translationsService,
+  )   : _permissionsService = permissionsService,
+        _translationsService = translationsService;
 
-  final PermissionsService permissionsService;
+  final PermissionsService _permissionsService;
+  final TranslationsService _translationsService;
+
   bool _appInitialized = false;
 
   Future<void> initializeApp() async {
-    await permissionsService.load();
+    await Future.wait(_nomenclatures);
+
     _appInitialized = true;
   }
+
+  List<Future<void>> get _nomenclatures => [
+        _permissionsService.load(),
+        _translationsService.load(),
+      ];
 
   bool get isAppInitialized => _appInitialized;
 }