+++ packages/rx_bloc_cli/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [3.3.0]
+* Added remote translation lookup for localizations
 * Updates to `analytics` feature - added crash reporting support and screen view logging
 
 ## [3.2.0]